+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/entrypoint/EventListener.java
@@ -3,14 +3,10 @@ package com.mageddo.dnsproxyserver.server.dns.solver.docker.entrypoint;
 import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.api.async.ResultCallback;
 import com.github.dockerjava.api.model.Event;
-import com.mageddo.dnsproxyserver.config.Configs;
 import com.mageddo.dnsproxyserver.di.StartupEvent;
-import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacade;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.Network;
-import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.ContainerService;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.DockerNetworkService;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.DpsDockerEnvironmentSetupService;
-import com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.DockerDAO;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -26,33 +22,16 @@ import java.io.Closeable;
 public class EventListener implements StartupEvent {
 
   private final DockerClient dockerClient;
-  private final DockerDAO dockerDAO;
   private final DpsDockerEnvironmentSetupService dpsDockerEnvironmentSetupService;
-  private final DockerNetworkFacade dockerNetworkDAO;
   private final DockerNetworkService networkService;
-  private final ContainerService containerService;
 
   @Override
   public void onStart() {
 
-    // tag:newMethod todo this code is responsability of an application service
-    final var dockerConnected = this.dockerDAO.isConnected();
-    log.info("status=binding-docker-events, dockerConnected={}", dockerConnected);
-    if (!dockerConnected) {
+    if (!this.dpsDockerEnvironmentSetupService.setup()) {
+      log.info("status=containerAutoConnectToDpsNetworkDisabled");
       return;
     }
-
-    this.dpsDockerEnvironmentSetupService.setup();
-    final var config = Configs.getInstance();
-    if (!config.getMustConfigureDpsNetwork() || !config.getDpsNetworkAutoConnect()) {
-      log.info(
-        "status=autoConnectDpsNetworkDisabled, dpsNetwork={}, dpsNetworkAutoConnect={}",
-        config.getMustConfigureDpsNetwork(), config.getDpsNetworkAutoConnect()
-      );
-      return;
-    }
-    this.containerService.connectRunningContainers();
-    // end:newMethod
 
     final var callback = new ResultCallback<Event>() {
       @Override
@@ -61,6 +40,7 @@ public class EventListener implements StartupEvent {
 
       @Override
       public void onStart(Closeable closeable) {
+        log.info("status=listeningContainersToConnectToDpsNetwork");
       }
 
       @Override
@@ -75,7 +55,7 @@ public class EventListener implements StartupEvent {
             return;
           }
           log.debug("status=eventIgnored, event={}", event);
-        } catch (Throwable e){
+        } catch (Throwable e) {
           log.warn("status=errorWhenProcessingEvent, msg={}, event={}", e.getMessage(), event, e);
         }
       }
@@ -90,7 +70,6 @@ public class EventListener implements StartupEvent {
     };
     this.dockerClient
       .eventsCmd()
-//      .withEventFilter("start", "die", "stop", "destroy")
       .withEventFilter("start")
       .exec(callback);
   }
